Add category validation and unescaping at the data-layer level
