+++ client/extensions/woocommerce/state/data-layer/product-categories/index.js
@@ -26,11 +26,10 @@ import {
 	WOOCOMMERCE_PRODUCT_CATEGORIES_REQUEST_SUCCESS,
 	WOOCOMMERCE_PRODUCT_CATEGORIES_REQUEST_FAILURE,
 } from 'woocommerce/state/action-types';
-import { dispatchRequest } from 'state/data-layer/wpcom-http/utils';
+import { dispatchRequestEx } from 'state/data-layer/wpcom-http/utils';
 import { DEFAULT_QUERY } from 'woocommerce/state/sites/product-categories/utils';
 import request from 'woocommerce/state/sites/http-request';
-
-// @TODO Move these handlers to product-categories/handlers.js
+import { verifyResponseHasValidCategories } from 'woocommerce/state/data-layer/utils';
 
 export function handleProductCategoryCreate( store, action ) {
 	const { siteId, category, successAction, failureAction } = action;
@@ -124,32 +123,17 @@ export function handleProductCategoryDeleteSuccess( { dispatch }, action, catego
 	} );
 }
 
-export function handleProductCategoriesRequest( { dispatch }, action ) {
+export const fetch = action => {
 	const { siteId, query } = action;
 	const requestQuery = { ...DEFAULT_QUERY, ...query };
 	const queryString = stringify( omitBy( requestQuery, val => '' === val ) );
 
-	dispatch( request( siteId, action ).getWithHeaders( `products/categories?${ queryString }` ) );
+	return request( siteId, action ).getWithHeaders( `products/categories?${ queryString }` );
 }
 
-export function handleProductCategoriesSuccess( { dispatch }, action, { data } ) {
+export const onFetchSuccess = ( action, { data } ) => dispatch => {
 	const { siteId, query } = action;
-	const { headers, body, status } = data;
-
-	// handleProductCategoriesRequest uses &_envelope
-	// ( https://developer.wordpress.org/rest-api/using-the-rest-api/global-parameters/#_envelope )
-	// so that we can get the X-WP-Total header back from the end-site. This means we will always get a 200
-	// status back, and the real status of the request will be stored in the response. This checks the real status.
-	if ( status !== 200 || ! isValidCategoriesArray( body ) ) {
-		dispatch( {
-			type: WOOCOMMERCE_PRODUCT_CATEGORIES_REQUEST_FAILURE,
-			siteId,
-			error: 'Invalid Categories Array',
-			query,
-		} );
-		dispatch( setError( siteId, action, { message: 'Invalid Categories Array', body } ) );
-		return;
-	}
+	const { headers, body } = data;
 
 	const total = headers[ 'X-WP-Total' ];
 	const totalPages = headers[ 'X-WP-TotalPages' ];
@@ -172,7 +156,7 @@ export function handleProductCategoriesSuccess( { dispatch }, action, { data } )
 	}
 }
 
-export function handleProductCategoriesError( { dispatch }, action, error ) {
+export const onFetchError = ( action, error ) => dispatch => {
 	const { siteId, query } = action;
 
 	dispatch( {
@@ -185,28 +169,6 @@ export function handleProductCategoriesError( { dispatch }, action, error ) {
 	dispatch( setError( siteId, action, error ) );
 }
 
-function isValidCategoriesArray( categories ) {
-	for ( let i = 0; i < categories.length; i++ ) {
-		if ( ! isValidProductCategory( categories[ i ] ) ) {
-			// Short-circuit the loop and return now.
-			return false;
-		}
-	}
-	return true;
-}
-
-function isValidProductCategory( category ) {
-	return (
-		category &&
-		category.id &&
-		'number' === typeof category.id &&
-		category.name &&
-		'string' === typeof category.name &&
-		category.slug &&
-		'string' === typeof category.slug
-	);
-}
-
 export default {
 	[ WOOCOMMERCE_PRODUCT_CATEGORY_CREATE ]: [ handleProductCategoryCreate ],
 	[ WOOCOMMERCE_PRODUCT_CATEGORY_DELETE ]: [
@@ -215,10 +177,11 @@ export default {
 	],
 	[ WOOCOMMERCE_PRODUCT_CATEGORY_UPDATE ]: [ handleProductCategoryUpdate ],
 	[ WOOCOMMERCE_PRODUCT_CATEGORIES_REQUEST ]: [
-		dispatchRequest(
-			handleProductCategoriesRequest,
-			handleProductCategoriesSuccess,
-			handleProductCategoriesError
-		),
+		dispatchRequestEx( {
+			fetch,
+			onSuccess: onFetchSuccess,
+			onError: onFetchError,
+			fromApi: verifyResponseHasValidCategories,
+		} ),
 	],
 };
